Staging

+++ public/scripts/extensions/regex/index.js
@@ -76,10 +76,27 @@ async function loadRegexScripts() {
         const scriptHtml = scriptTemplate.clone();
         scriptHtml.attr('id', uuidv4());
         scriptHtml.find('.regex_script_name').text(script.scriptName);
-        scriptHtml.find('.edit_existing_regex').on('click', async function() {
+        scriptHtml.find('.disable_regex').prop("checked", script.disabled ?? false)
+            .on('input', function () {
+                script.disabled = !!$(this).prop("checked");
+                saveSettingsDebounced();
+            });
+        scriptHtml.find('.regex-toggle-on').on('click', function () {
+            scriptHtml.find('.disable_regex').prop("checked", true).trigger('input');
+        });
+        scriptHtml.find('.regex-toggle-off').on('click', function () {
+            scriptHtml.find('.disable_regex').prop("checked", false).trigger('input');
+        });
+        scriptHtml.find('.edit_existing_regex').on('click', async function () {
             await onRegexEditorOpenClick(scriptHtml.attr("id"));
         });
-        scriptHtml.find('.delete_regex').on('click', async function() {
+        scriptHtml.find('.delete_regex').on('click', async function () {
+            const confirm = await callPopup("Are you sure you want to delete this regex script?", "confirm");
+
+            if (!confirm) {
+                return;
+            }
+
             await deleteRegexScript({ existingId: scriptHtml.attr("id") });
         });
 
@@ -113,6 +130,9 @@ async function onRegexEditorOpenClick(existingId) {
             editorHtml
                 .find(`input[name="only_format_display"]`)
                 .prop("checked", existingScript.markdownOnly ?? false);
+            editorHtml
+                .find(`input[name="only_format_prompt"]`)
+                .prop("checked", existingScript.promptOnly ?? false);
             editorHtml
                 .find(`input[name="run_on_edit"]`)
                 .prop("checked", existingScript.runOnEdit ?? false);
@@ -154,7 +174,7 @@ async function onRegexEditorOpenClick(existingId) {
                 editorHtml
                     .find(`input[name="replace_position"]`)
                     .filter(":checked")
-                    .map(function() { return parseInt($(this).val()) })
+                    .map(function () { return parseInt($(this).val()) })
                     .get()
                     .filter((e) => e !== NaN) || [],
             disabled:
@@ -164,6 +184,10 @@ async function onRegexEditorOpenClick(existingId) {
             markdownOnly:
                 editorHtml
                     .find(`input[name="only_format_display"]`)
+                    .prop("checked"),
+            promptOnly:
+                editorHtml
+                    .find(`input[name="only_format_prompt"]`)
                     .prop("checked"),
             runOnEdit:
                 editorHtml
@@ -197,6 +221,7 @@ function migrateSettings() {
                 script.placement = script.placement.filter((e) => e !== regex_placement.MD_DISPLAY);
 
             script.markdownOnly = true
+            script.promptOnly = true
 
             performSave = true;
         }
@@ -231,7 +256,7 @@ jQuery(async () => {
 
     const settingsHtml = await $.get("scripts/extensions/regex/dropdown.html");
     $("#extensions_settings2").append(settingsHtml);
-    $("#open_regex_editor").on("click", function() {
+    $("#open_regex_editor").on("click", function () {
         onRegexEditorOpenClick(false);
     });
